More swagger encapsulation
